requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SocketIO==5.6.1
librosa==0.10.1
torch==2.1.0
torchaudio==2.1.0
numpy==1.26.0
werkzeug==3.0.0
requests==2.31.0
scikit-learn==1.3.1
soundfile==0.12.1
matplotlib==3.8.0
python-dotenv==1.0.1
SpeechRecognition==3.10.4
groq==0.11.0
imageio-ffmpeg==0.5.1
edge-tts==6.1.12
gunicorn==23.0.0
simple-websocket==1.1.0
sounddevice==0.4.6
webrtcvad-wheels==2.0.14
setuptools>=69.0.0,<81.0.0
